+++ pyvo/registry/tests/commonfixtures.py
@@ -6,6 +6,7 @@ Common fixtures for pyVO registry tests
 import pytest
 
 from astropy.utils.data import (
+    clear_download_cache,
     get_pkg_data_filename,
     import_file_to_cache)
 # We need to populate the vocabulary cache with our test data;
@@ -24,6 +25,44 @@ def messenger_vocabulary(mocker):
         get_pkg_data_filename(
             'data/messenger.desise',
             package=__package__))
+
+
+# The moc UAT was produced by this program:
+# import json
+# import pyvo
+#
+# def gather_children(voc, t):
+# 	result = {t}
+# 	voc[t].pop("description", None)
+# 	for c in voc[t]["narrower"]:
+# 		result |= gather_children(voc, c)
+# 	return result
+#
+# uat = pyvo.utils.vocabularies.get_vocabulary("uat")
+# solphys = gather_children(uat["terms"], "solar-physics")
+# uat["terms"] = {t: m for t, m in uat["terms"].items()
+# 	if t in solphys}
+# with open("uat-selection.desise", "w", encoding="utf-8") as f:
+# 	json.dump(uat, f, indent=1)
+
+@pytest.fixture()
+def uat_vocabulary(mocker):
+    """a small sample of the IVOA UAT vocabulary in astropy's cache.
+
+    We need to clean up behind ourselves, because our version of the
+    UAT is limited to the solar-physics branch in order to not waste
+    too much space.  The source code here contains a program to refresh
+    this vocabulary selection.
+    """
+    import_file_to_cache(
+        'http://www.ivoa.net/rdf/uat',
+        get_pkg_data_filename(
+            'data/uat-selection.desise',
+            package=__package__))
+    yield
+    # it would be nice if we only did that if we polluted the
+    # cache before the yield, but we can't easily see if we did that.
+    clear_download_cache('http://www.ivoa.net/rdf/uat')
 
 
 # We need an object standing in for TAP services for query generation.