+++ CHANGES.rst
@@ -10,6 +10,10 @@ Enhancements and Fixes
 
 - Change AsyncTAPJob.result to return None if no result is found explicitly [#644]
 
+- Add a UAT constraint to the registry interface for constraining
+  subjects [#XXX]
+
+
 
 Deprecations and Removals
 -------------------------